backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
setuptools==69.5.1
sqlparse==0.5.1
wheel==0.43.0
